Controller Integration Testing Complete
