+++ backend/src/main/resources/application.properties
@@ -1,3 +1,3 @@
-spring.profiles.active = test
+spring.profiles.active = aws
