+++ backend/src/test/java/com/qa/citizen/rest/WhereaboutsDTOControllerIntegrationTest.java
@@ -19,10 +19,10 @@ import org.springframework.test.web.servlet.RequestBuilder;
 import org.springframework.test.web.servlet.ResultMatcher;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.qa.citizen.domain.Citizen;
 import com.qa.citizen.rest.DTOs.LocationAtmDTO;
 import com.qa.citizen.rest.DTOs.LocationEposDTO;
 import com.qa.citizen.rest.DTOs.WhereaboutsDTO;
+import com.qa.citizen.service.IntegrationTestService;
 
 @SpringBootTest(webEnvironment = WebEnvironment.RANDOM_PORT) // loads the context
 @AutoConfigureMockMvc
@@ -35,20 +35,8 @@ public class WhereaboutsDTOControllerIntegrationTest {
 	@Autowired
 	private ObjectMapper mapper;
 
-	public String createReceivedCitizenAsJSON() throws Exception {
-
-		// get passed something close to a citizen object
-		// create passed Citizen
-		Citizen passedCitizen = new Citizen();
-		passedCitizen.setForenames("Michael Shane");
-		passedCitizen.setSurname("Cochrane");
-
-		// convert passed citizen object to json string
-		String passedCitizenAsJSON = this.mapper.writeValueAsString(passedCitizen);
-
-		return passedCitizenAsJSON;
-
-	}
+	@Autowired
+	private IntegrationTestService service;
 
 	// WhereaboutsDTOControllerIntegrationTest
 	@Test
@@ -56,7 +44,7 @@ public class WhereaboutsDTOControllerIntegrationTest {
 
 		// build a mock request
 		RequestBuilder mockRequest = post("/getWhereabouts/").contentType(MediaType.APPLICATION_JSON)
-				.content(createReceivedCitizenAsJSON());
+				.content(service.createReceivedCitizenAsJSON());
 
 		// return a list of WhereaboutsDTO objects that match the citizen object
 		List<WhereaboutsDTO> whereaboutsDTOList = new ArrayList<>();
